% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_arqtra.R
\name{import_arqtra}
\alias{import_arqtra}
\title{Importar arquivo de transações monetárias de cartões Visa}
\usage{
import_arqtra(
  caminho,
  mysql = FALSE,
  schema = "ARQTRA",
  tabela,
  drop = FALSE,
  metodo = c("vroom", "readr"),
  host,
  port,
  user,
  password = NULL
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensão (.txt).}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que conterá a tabela
ARQTRA no MySQL. O schema deve existir no MySQL antes de executar a função.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que conterá os dados importados.}

\item{MySQL}{Logical. Se TRUE, entao o arquivo será importado para o MySQL.}
}
\description{
Importa o arquivo ARQTRA para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da função. Caso for importar para o MySQL, não é necessário
criar um objeto. Note que a tabela será sobrescrita se já existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_gco("C:\\\\Arquivos\\\\GCO\\\\202009.txt")

# para importar para o MySQL:
import_gco(
    "C:\\\\Arquivos\\\\GCO\\\\202009.txt",
    mysql = TRUE,
    tabela = "t202009",
    user = "alberson"
)
}

}
