% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_BC.R
\name{import_BC}
\alias{import_BC}
\title{Importar arquivo de impedimentos da GBR}
\usage{
import_BC(caminho, MySQL = FALSE, schema = "GBR", tabela = "BC")
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que contera a tabela
BC no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}
}
\description{
Importa o arquivo BC de impedimentos da GBR para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_BC("C:\\\\Arquivos\\\\GBR\\\\202009.txt")

# para importar para o MySQL:
import_BC("C:\\\\Arquivos\\\\GBR\\\\202009.txt", MySQL = TRUE)
# ou se nao quiser sobrescrever a tabela:
import_BC("C:\\\\Arquivos\\\\GBR\\\\202009.txt", MySQL = TRUE, tabela = "202009")
}

}
