% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bg.R
\name{import_BG}
\alias{import_BG}
\title{Importar arquivo da base geral de clientes do GCB}
\usage{
import_BG(caminho, tipo, MySQL = FALSE, schema = "gcb", tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{tipo}{String. pf para pessoa fisica e pj para pessoa juridica.}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que contera a tabela
BG no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}
}
\description{
Importa o arquivo BG para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_BG("C:\\\\Arquivos\\\\GCB\\\\IF041020.txt", "pf")

# para importar para o MySQL:
import_GCO("C:\\\\Arquivos\\\\GCB\\\\IF041020.txt", "pf", MySQL = TRUE, tabela = "bg")
}

}
