% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_C3.R
\name{import_c3}
\alias{import_c3}
\title{Importar arquivo de PNs da gestão de crédito}
\usage{
import_c3(caminho, mysql = FALSE, schema = "c5", tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que conterá a tabela
MG5 no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}
}
\description{
Importa o arquivo C3 para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao é necessário
criar um objeto. Note que a tabela será sobrescrita se já existir.
}
\examples{
\dontrun{
# para importar para o R:
data <- import_c3
3("C:\\\\Arquivos\\\\C3\\\\202009.txt")

# para importar para o MySQL:
import_c3("C:\\\\Arquivos\\\\C3\\\\202009.txt", MySQL = TRUE, tabela = "t202009")
}

}
