% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_got.R
\name{import_got}
\alias{import_got}
\title{Importar arquivo de unidades do SFB}
\usage{
import_got(caminho, MySQL = FALSE, schema = "GOT", tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que contera a tabela
GCO no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}
}
\description{
Importa o arquivo GOT78 para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_GCO("C:\\\\Arquivos\\\\GCO\\\\202009.txt")

# para importar para o MySQL:
import_GCO("C:\\\\Arquivos\\\\GCO\\\\202009.txt", MySQL = TRUE, tabela = "t202009")
}

}
