% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_MG5.R
\name{import_mg5}
\alias{import_mg5}
\title{Importar arquivo de renegociação da gestão de crédito}
\usage{
import_mg5(caminho, MySQL = FALSE, schema = "mg5", tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que conterá a tabela
MG5 no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}
}
\description{
Importa o arquivo GCOA53 para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao é necessário
criar um objeto. Note que a tabela será sobrescrita se já existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_MG5("C:\\\\Arquivos\\\\MG5\\\\202009.txt")

# para importar para o MySQL:
import_MG5("C:\\\\Arquivos\\\\MG5\\\\202009.txt", MySQL = TRUE, tabela = "t202009")
}

}
