% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_SCR.R
\name{import_SCR}
\alias{import_SCR}
\title{Importar arquivo de operacoes de credito do SCR/BACEN}
\usage{
import_SCR(caminho, MySQL = FALSE, schema = "SCR", tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que contera a tabela
GCO no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}
}
\description{
Importa o arquivo SCR do Banco Central do Brasil com todas as operacoes
de instituicoes financeiras dos clientes Banestes para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_SCR("C:\\\\Arquivos\\\\SCR\\\\202009.txt")

# para importar para o MySQL:
import_BC("C:\\\\Arquivos\\\\SCR\\\\202009.txt", MySQL = TRUE, tabela = "t202009")
}

}
