% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_MySQL.R
\name{send_MySQL}
\alias{send_MySQL}
\title{Envia objeto do R para o MySQL}
\usage{
send_MySQL(objeto, schema, tabela)
}
\arguments{
\item{objeto}{O a ser enviado ao MySQL.}

\item{schema}{String. O nome do schema que contera a tabela.}

\item{tabela}{String. O nome da tabela a ser criada no
servidor que contera os dados importados.}
}
\description{
Envia um objeto no ambiente do R para o MySQL. Recomendado quando nao ha
memoria suficiente para o envio direto ao MySQL.
}
\details{
Quando um arquivo importado e muito grande, como o SCR, pode nao
haver memoria disponivel para enviar o arquivo ao MySQL. Nesse caso, importe
ao R, libere memoria e entao envie ao MySQL. Note que a tabela sera
sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_SCR("C:\\\\Arquivos\\\\SCR\\\\202009.txt")

# para enviar ao MySQL:
send_MySQL(data, schema = "SCR", tabela = "t202009")
}

}
