% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gco.R
\name{import_gco}
\alias{import_gco}
\title{Importar arquivo de contratos de credito do sistema GCO}
\usage{
import_gco(
  caminho,
  mysql = FALSE,
  schema = "GCO",
  tabela,
  metodo = c("vroom", "readr"),
  host,
  user
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{schema}{String. Se MySQL = TRUE, o nome do schema que contera a tabela
GCO no MySQL. O schema deve existir no MySQL antes de executar a funcao.}

\item{tabela}{String. Se MySQL = TRUE, o nome da tabela a ser criada no
servidor que contera os dados importados.}

\item{MySQL}{Logical. Se TRUE, entao o arquivo sera importado para o MySQL.}
}
\description{
Importa o arquivo GCOA53 para o R ou para o MySQL.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o MySQL, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data <- import_gco("C:\\\\Arquivos\\\\GCO\\\\202009.txt")

# para importar para o MySQL:
import_gco(
  "C:\\\\Arquivos\\\\GCO\\\\202009.txt",
  mysql = TRUE,
  tabela = "t202009",
  user = "alberson"
)
}

}
